backend/requirements.txt
asgiref==3.8.1
channels==4.1.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
